requirements.txt
amqp==5.0.6
anyjson==0.3.3
arrow==1.1.0
asgiref==3.3.4
attrs==21.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
blessed==1.18.0
cached-property==1.5.2
celery==5.1.0
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
dataclasses==0.8
Django==3.2.4
django-celery-results==2.0.1
django-filter==2.4.0
django-picklefield==3.0.1
django-q==1.3.8
djangorestframework==3.12.4
idna==2.10
importlib-metadata==4.5.0
iniconfig==1.1.1
iso4217parse==0.5.1
kombu==5.1.0
lxml==4.6.3
Markdown==3.3.4
packaging==20.9
pluggy==0.13.1
prompt-toolkit==3.0.18
py==1.10.0
pymongo==3.11.4
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.2.4
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.1
